# Conflicts:
#	src/main/java/nl/miwgroningen/se6/heartcoded/CaTo/controller/MemberController.java
#	src/main/resources/templates/groupUpdateMember.html

+++ src/main/java/nl/miwgroningen/se6/heartcoded/CaTo/controller/UserController.java
@@ -2,6 +2,7 @@ package nl.miwgroningen.se6.heartcoded.CaTo.controller;
 
 import nl.miwgroningen.se6.heartcoded.CaTo.dto.UserDTO;
 import nl.miwgroningen.se6.heartcoded.CaTo.dto.UserRegistrationDTO;
+import nl.miwgroningen.se6.heartcoded.CaTo.service.MemberService;
 import nl.miwgroningen.se6.heartcoded.CaTo.service.TaskListService;
 import nl.miwgroningen.se6.heartcoded.CaTo.service.UserService;
 import org.springframework.stereotype.Controller;
@@ -12,6 +13,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 /**
  * @author Shalena Omapersad <shalenao@hotmail.com>
@@ -22,10 +24,15 @@ import org.springframework.web.bind.annotation.PostMapping;
 @Controller
 public class UserController {
 
-    private UserService userService;
+    private static final String USER_IS_THE_LAST_ADMIN_IN_A_GROUP = "User is the last admin in a group";
+    private static final String SITE_ADMIN_DELETE_ERROR = "Unable to delete site admin";
 
-    public UserController(UserService userService) {
+    private UserService userService;
+    private MemberService memberService;
+
+    public UserController(UserService userService, MemberService memberService) {
         this.userService = userService;
+        this.memberService = memberService;
     }
 
     @GetMapping("/registration")
@@ -36,13 +43,25 @@ public class UserController {
     }
 
     @GetMapping("/users/delete/{userId}")
-    protected String deleteUser(@PathVariable("userId") Integer userId) {
-        userService.deleteUserById(userId);
+    protected String deleteUser(@PathVariable("userId") Integer userId, RedirectAttributes redirectAttributes) {
+        if (!userService.currentUserIsSiteAdmin()) {
+            return "redirect:/403";
+        }
+        if (memberService.userIsLastGroupAdminInAnyGroup(userId)) {
+            redirectAttributes.addAttribute("error", USER_IS_THE_LAST_ADMIN_IN_A_GROUP);
+        } else if (userService.userIsSiteAdmin(userId)) {
+            redirectAttributes.addAttribute("error", SITE_ADMIN_DELETE_ERROR);
+        } else {
+            userService.deleteUserById(userId);
+        }
         return "redirect:/siteAdminDashboard";
     }
 
     @GetMapping("/users/edit/{userId}")
     protected String showEditUserForm(@PathVariable("userId") Integer userId, Model model) {
+        if (!userService.getCurrentUser().getUserId().equals(userId)) {
+            return "redirect:/403";
+        }
         UserDTO user = userService.getById(userId);
         model.addAttribute("user", user);
         return "editUserForm";
@@ -66,6 +85,10 @@ public class UserController {
     protected String editUser(@ModelAttribute("user") UserDTO user,
                               BindingResult result) {
         UserDTO currentUser = userService.getCurrentUser();
+        if (!userService.getCurrentUser().getUserId().equals(user.getUserId()) ||
+                userService.currentUserIsSiteAdmin()) {
+            return "redirect:/403";
+        }
         if (userService.emailInUse(user.getEmail()) && !user.getEmail().equals(currentUser.getEmail())) {
             ObjectError error = new ObjectError("globalError", "Email is already in use");
             result.addError(error);
